+++ server/user-service/src/main/java/org/wid/userservice/adapter/primary/rest/controller/UserController.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.wid.userservice.dto.user.UserDto;
-import org.wid.userservice.entity.entity.User.LoginType;
+import org.wid.userservice.entity.User.LoginType;
 import org.wid.userservice.port.primary.UserServicePort;
 
 import lombok.RequiredArgsConstructor;
