+++ server/user-service/src/main/java/org/wid/userservice/dto/auth/Oauth2LoginRequestDto.java
@@ -1,6 +1,6 @@
 package org.wid.userservice.dto.auth;
 
-import org.wid.userservice.entity.entity.User.LoginType;
+import org.wid.userservice.entity.User.LoginType;
 
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
