+++ server/user-service/src/main/java/org/wid/userservice/entity/User.java
@@ -1,4 +1,4 @@
-package org.wid.userservice.entity.entity;
+package org.wid.userservice.entity;
 
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
