+++ server/user-service/src/main/java/org/wid/userservice/service/UserService.java
@@ -2,7 +2,7 @@ package org.wid.userservice.service;
 
 import org.springframework.stereotype.Service;
 import org.wid.userservice.dto.user.UserDto;
-import org.wid.userservice.entity.entity.User;
+import org.wid.userservice.entity.User;
 import org.wid.userservice.mapper.UserMapper;
 import org.wid.userservice.port.primary.UserServicePort;
 import org.wid.userservice.port.secondary.UserRepositoryPort;
